+++ src/main/simplejpa/scaffolding/generator/basic/DateAttributeGenerator.groovy
@@ -49,4 +49,14 @@ class DateAttributeGenerator extends BuiltInAttributeGenerator {
         return ["model.$name = ${var}.$name"]
     }
 
+    @Override
+    List<String> action() {
+        []
+    }
+
+    @Override
+    List<String> popup() {
+        []
+    }
+
 }