+++ src/main/simplejpa/scaffolding/generator/basic/UnknownAttributeGenerator.groovy
@@ -61,4 +61,14 @@ class UnknownAttributeGenerator extends BuiltInAttributeGenerator {
         ]
     }
 
+    @Override
+    List<String> action() {
+        []
+    }
+
+    @Override
+    List<String> popup() {
+        []
+    }
+
 }