+++ src/templates/artifacts/StartupController.groovy
@@ -29,9 +29,11 @@ class ${g.scaffolding.startupGroupName}Controller {
 		def (m,v,c) = app.mvcGroupManager.createMVCGroup(groupId, groupId)
 
 		execInsideUIAsync {
-			view.mainPanel.add(v.mainPanel, groupId)
-			view.cardLayout.show(view.mainPanel, groupId)
-			view.mainFrame.title = "\${app.config.application.title}: \${GriffonNameUtils.getNaturalName(groupId)}"
+            view.mainPanel.removeAll()
+            view.mainPanel.add(v.mainPanel, BorderLayout.CENTER)
+            view.mainPanel.revalidate()
+            view.mainPanel.repaint()
+			view.mainFrame.title = "\${app.config.application.title} \${app.metadata.getApplicationVersion()}: \${GriffonNameUtils.getNaturalName(groupId)}"
 			view.busyLabel.visible = false
 		}
 	}